refactor exercise 15.9
